Refactor: introduce EdgeRendering
