+++ src/main/java/com/powsybl/nad/svg/SvgWriter.java
@@ -20,7 +20,6 @@ import java.nio.file.Path;
 import java.util.*;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
-import java.util.stream.IntStream;
 
 /**
  * @author Florian Dupuy <florian.dupuy at rte-france.com>
@@ -53,11 +52,13 @@ public class SvgWriter {
     private final SvgParameters svgParameters;
     private final StyleProvider styleProvider;
     private final LabelProvider labelProvider;
+    private final EdgeRendering edgeRendering;
 
     public SvgWriter(SvgParameters svgParameters, StyleProvider styleProvider, LabelProvider labelProvider) {
         this.svgParameters = Objects.requireNonNull(svgParameters);
         this.styleProvider = Objects.requireNonNull(styleProvider);
         this.labelProvider = Objects.requireNonNull(labelProvider);
+        this.edgeRendering = new DefaultEdgeRendering();
     }
 
     public void writeSvg(Graph graph, Path svgFile) {
@@ -87,7 +88,7 @@ public class SvgWriter {
         Objects.requireNonNull(svgOs);
 
         // Edge coordinates need to be computed first, based on svg parameters
-        computeBranchEdgesCoordinates(graph);
+        edgeRendering.run(graph, svgParameters);
 
         try {
             XMLStreamWriter writer = XmlUtil.initializeWriter(true, INDENT, svgOs);
@@ -104,205 +105,6 @@ public class SvgWriter {
         } catch (XMLStreamException e) {
             throw new UncheckedXmlStreamException(e);
         }
-    }
-
-    private void computeBranchEdgesCoordinates(Graph graph) {
-        graph.getNonMultiBranchEdgesStream().forEach(edge -> computeSingleBranchEdgeCoordinates(graph, edge));
-        graph.getMultiBranchEdgesStream().forEach(edges -> computeMultiBranchEdgesCoordinates(graph, edges));
-        graph.getLoopBranchEdgesMap().forEach((node, edges) -> loopEdgesLayout(graph, node, edges));
-        graph.getThreeWtEdgesStream().forEach(edge -> computeThreeWtEdgeCoordinates(graph, edge));
-        graph.getTextEdgesMap().forEach((edge, nodes) -> computeTextEdgeLayoutCoordinates(nodes.getFirst(), nodes.getSecond(), edge));
-    }
-
-    private void computeTextEdgeLayoutCoordinates(Node node1, Node node2, TextEdge edge) {
-        edge.setPoints(node1.getPosition(), node2.getPosition());
-    }
-
-    private void computeSingleBranchEdgeCoordinates(Graph graph, BranchEdge edge) {
-        Node node1 = graph.getBusGraphNode1(edge);
-        Node node2 = graph.getBusGraphNode2(edge);
-
-        Point direction1 = getDirection(node2, () -> graph.getNode2(edge));
-        Point edgeStart1 = computeEdgeStart(node1, direction1, () -> graph.getNode1(edge));
-
-        Point direction2 = getDirection(node1, () -> graph.getNode1(edge));
-        Point edgeStart2 = computeEdgeStart(node2, direction2, () -> graph.getNode2(edge));
-
-        Point middle = Point.createMiddlePoint(edgeStart1, edgeStart2);
-        edge.setPoints1(edgeStart1, middle);
-        edge.setPoints2(edgeStart2, middle);
-    }
-
-    private Point getDirection(Node directionBusGraphNode, Supplier<Node> vlNodeSupplier) {
-        if (directionBusGraphNode == BusNode.UNKNOWN) {
-            return vlNodeSupplier.get().getPosition();
-        }
-        return directionBusGraphNode.getPosition();
-    }
-
-    private Point computeEdgeStart(Node node, Point direction, Supplier<Node> vlNodeSupplier) {
-        // If edge not connected to a bus node on that side, we use corresponding voltage level with specific extra radius
-        if (node == BusNode.UNKNOWN) {
-            double unknownBusRadius = svgParameters.getVoltageLevelCircleRadius() + svgParameters.getUnknownBusNodeExtraRadius();
-            return vlNodeSupplier.get().getPosition().atDistance(unknownBusRadius, direction);
-        }
-
-        Point edgeStart = node.getPosition();
-        if (node instanceof BusNode) {
-            int nbNeighbours = ((BusNode) node).getNbNeighbouringBusNodes();
-            double unitaryRadius = svgParameters.getVoltageLevelCircleRadius() / (nbNeighbours + 1);
-            double busAnnulusOuterRadius = (((BusNode) node).getIndex() + 1) * unitaryRadius - svgParameters.getEdgeStartShift();
-            edgeStart = edgeStart.atDistance(busAnnulusOuterRadius, direction);
-        }
-        return edgeStart;
-    }
-
-    private void computeMultiBranchEdgesCoordinates(Graph graph, List<BranchEdge> edges) {
-        Edge firstEdge = edges.iterator().next();
-        Node nodeA = graph.getNode1(firstEdge);
-        Node nodeB = graph.getNode2(firstEdge);
-        Point pointA = nodeA.getPosition();
-        Point pointB = nodeB.getPosition();
-
-        double dx = pointB.getX() - pointA.getX();
-        double dy = pointB.getY() - pointA.getY();
-        double angle = Math.atan2(dy, dx);
-
-        int nbForks = edges.size();
-        double forkAperture = svgParameters.getEdgesForkAperture();
-        double forkLength = svgParameters.getEdgesForkLength();
-        double angleStep = forkAperture / (nbForks - 1);
-
-        int i = 0;
-        for (BranchEdge edge : edges) {
-            if (2 * i + 1 == nbForks) { // in the middle, hence alpha = 0
-                computeSingleBranchEdgeCoordinates(graph, edge);
-            } else {
-                double alpha = -forkAperture / 2 + i * angleStep;
-                double angleForkA = angle - alpha;
-                double angleForkB = angle + Math.PI + alpha;
-                Point forkA = pointA.shift(forkLength * Math.cos(angleForkA), forkLength * Math.sin(angleForkA));
-                Point forkB = pointB.shift(forkLength * Math.cos(angleForkB), forkLength * Math.sin(angleForkB));
-                Point middle = Point.createMiddlePoint(forkA, forkB);
-
-                BranchEdge.Side sideA = graph.getNode1(edge) == nodeA ? BranchEdge.Side.ONE : BranchEdge.Side.TWO;
-                BranchEdge.Side sideB = sideA.getOpposite();
-
-                Node busNodeA = sideA == BranchEdge.Side.ONE ? graph.getBusGraphNode1(edge) : graph.getBusGraphNode2(edge);
-                Node busNodeB = sideA == BranchEdge.Side.ONE ? graph.getBusGraphNode2(edge) : graph.getBusGraphNode1(edge);
-
-                Point edgeStartA = computeEdgeStart(busNodeA, forkA, () -> nodeA);
-                edge.setPoints(sideA, edgeStartA, forkA, middle);
-
-                Point edgeStartB = computeEdgeStart(busNodeB, forkB, () -> nodeB);
-                edge.setPoints(sideB, edgeStartB, forkB, middle);
-            }
-            i++;
-        }
-    }
-
-    private void loopEdgesLayout(Graph graph, Node node, List<BranchEdge> loopEdges) {
-        List<Double> angles = computeLoopAngles(graph, loopEdges, node);
-
-        int i = 0;
-        Point nodePoint = node.getPosition();
-        for (BranchEdge edge : loopEdges) {
-            double angle = angles.get(i++);
-            Point middle = nodePoint.atDistance(svgParameters.getLoopDistance(), angle);
-            Point fork1 = nodePoint.atDistance(svgParameters.getEdgesForkLength(), angle - svgParameters.getLoopEdgesAperture() / 2);
-            Point fork2 = nodePoint.atDistance(svgParameters.getEdgesForkLength(), angle + svgParameters.getLoopEdgesAperture() / 2);
-
-            Node busNode1 = graph.getBusGraphNode1(edge);
-            Node busNode2 = graph.getBusGraphNode2(edge);
-
-            Point edgeStart1 = computeEdgeStart(busNode1, fork1, () -> node);
-            edge.setPoints(BranchEdge.Side.ONE, edgeStart1, fork1, middle);
-
-            Point edgeStart2 = computeEdgeStart(busNode2, fork2, () -> node);
-            edge.setPoints(BranchEdge.Side.TWO, edgeStart2, fork2, middle);
-        }
-    }
-
-    private List<Double> computeLoopAngles(Graph graph, List<BranchEdge> loopEdges, Node node) {
-        int nbLoops = loopEdges.size();
-
-        List<Double> anglesOtherEdges = graph.getBranchEdgeStream(node)
-                .filter(e -> !loopEdges.contains(e))
-                .mapToDouble(e -> getAngle(e, graph, node))
-                .sorted().boxed().collect(Collectors.toList());
-
-        List<Double> loopAngles = new ArrayList<>();
-        if (anglesOtherEdges.size() > 0) {
-            anglesOtherEdges.add(anglesOtherEdges.get(0) + 2 * Math.PI);
-            double apertureWithMargin = svgParameters.getLoopEdgesAperture() * 1.2;
-
-            double[] deltaAngles = new double[anglesOtherEdges.size() - 1];
-            int nbSeparatedSlots = 0;
-            int nbSharedSlots = 0;
-            for (int i = 0; i < anglesOtherEdges.size() - 1; i++) {
-                deltaAngles[i] = anglesOtherEdges.get(i + 1) - anglesOtherEdges.get(i);
-                nbSeparatedSlots += deltaAngles[i] > apertureWithMargin ? 1 : 0;
-                nbSharedSlots += Math.floor(deltaAngles[i] / apertureWithMargin);
-            }
-
-            List<Integer> sortedIndices = IntStream.range(0, deltaAngles.length)
-                    .boxed().sorted(Comparator.comparingDouble(i -> deltaAngles[i]))
-                    .collect(Collectors.toList());
-
-            if (nbLoops <= nbSeparatedSlots) {
-                // Place loops in "slots" separated by non-loop edges
-                for (int i = sortedIndices.size() - nbLoops; i < sortedIndices.size(); i++) {
-                    int iSorted = sortedIndices.get(i);
-                    loopAngles.add((anglesOtherEdges.get(iSorted) + anglesOtherEdges.get(iSorted + 1)) / 2);
-                }
-            } else if (nbLoops <= nbSharedSlots) {
-                // Place the maximum of loops in "slots" separated by non-loop edges, and put the excessive ones in the bigger "slots"
-                int nbExcessiveRemaining = nbLoops - nbSeparatedSlots;
-                for (int i = sortedIndices.size() - 1; i >= 0; i--) {
-                    int iSorted = sortedIndices.get(i);
-                    int nbAvailableSlots = (int) Math.floor(deltaAngles[iSorted] / apertureWithMargin);
-                    if (nbAvailableSlots == 0) {
-                        break;
-                    }
-                    int nbLoopsInDelta = Math.min(nbAvailableSlots, nbExcessiveRemaining + 1);
-                    double extraSpace = deltaAngles[iSorted] - svgParameters.getLoopEdgesAperture() * nbLoopsInDelta; // extra space without margins
-                    double intraSpace = extraSpace / (nbLoopsInDelta + 1); // space between two loops and between non-loop edges and first/last loop
-                    double angleStep = (anglesOtherEdges.get(iSorted + 1) - anglesOtherEdges.get(iSorted) - intraSpace) / nbLoopsInDelta;
-                    double startAngle = anglesOtherEdges.get(iSorted) + intraSpace / 2 + angleStep / 2;
-                    IntStream.range(0, nbLoopsInDelta).mapToDouble(iLoop -> startAngle + iLoop * angleStep).forEach(loopAngles::add);
-                    nbExcessiveRemaining -= nbLoopsInDelta - 1;
-                }
-            } else {
-                // Not enough place in the slots: dividing the circle in nbLoops, starting in the middle of the biggest slot
-                int iMaxDelta = sortedIndices.get(sortedIndices.size() - 1);
-                double startAngle = (anglesOtherEdges.get(iMaxDelta) + anglesOtherEdges.get(iMaxDelta + 1)) / 2;
-                IntStream.range(0, nbLoops).mapToDouble(i -> startAngle + i * 2 * Math.PI / nbLoops).forEach(loopAngles::add);
-            }
-
-        } else {
-            // No other edges: dividing the circle in nbLoops
-            IntStream.range(0, nbLoops).mapToDouble(i -> i * 2 * Math.PI / nbLoops).forEach(loopAngles::add);
-        }
-
-        return loopAngles;
-    }
-
-    private double getAngle(BranchEdge edge, Graph graph, Node node) {
-        BranchEdge.Side side = graph.getNode1(edge) == node ? BranchEdge.Side.ONE : BranchEdge.Side.TWO;
-        return getEdgeStartAngle(edge.getPoints(side));
-    }
-
-    private void computeThreeWtEdgeCoordinates(Graph graph, ThreeWtEdge edge) {
-        Node node1 = graph.getBusGraphNode1(edge);
-        Node node2 = graph.getBusGraphNode2(edge);
-
-        Point direction1 = getDirection(node2, () -> graph.getNode2(edge));
-        Point edgeStart1 = computeEdgeStart(node1, direction1, () -> graph.getNode1(edge));
-
-        Point direction2 = getDirection(node1, () -> graph.getNode1(edge));
-        Point edgeStart2 = computeEdgeStart(node2, direction2, () -> graph.getNode2(edge));
-
-        edge.setPoints(edgeStart1, edgeStart2);
     }
 
     private void drawBranchEdges(Graph graph, XMLStreamWriter writer) throws XMLStreamException {
@@ -355,9 +157,6 @@ public class SvgWriter {
         writer.writeStartElement(GROUP_ELEMENT_NAME);
         writer.writeAttribute(CLASS_ATTRIBUTE, StyleProvider.THREE_WT_EDGES_CLASS);
         for (ThreeWtEdge edge : threeWtEdges) {
-            if (!edge.isVisible()) {
-                continue;
-            }
             drawThreeWtEdge(graph, writer, edge);
         }
         writer.writeEndElement();
@@ -370,30 +169,40 @@ public class SvgWriter {
         }
         writer.writeStartElement(GROUP_ELEMENT_NAME);
         addStylesIfAny(writer, styleProvider.getSideEdgeStyleClasses(edge, side));
-        List<Point> half = edge.getPoints(side);
         if (edge.isVisible(side)) {
             if (!graph.isLoop(edge)) {
                 writer.writeEmptyElement(POLYLINE_ELEMENT_NAME);
-                String lineFormatted = half.stream()
-                        .map(point -> getFormattedValue(point.getX()) + "," + getFormattedValue(point.getY()))
-                        .collect(Collectors.joining(" "));
-                writer.writeAttribute(POINTS_ATTRIBUTE, lineFormatted);
-                drawEdgeInfo(writer, half, labelProvider.getEdgeInfos(graph, edge, side));
+                writer.writeAttribute(POINTS_ATTRIBUTE, getPolylinePointsString(edge, side));
+                drawBranchEdgeInfo(writer, edge, side, labelProvider.getEdgeInfos(graph, edge, side));
             } else {
                 writer.writeEmptyElement(PATH_ELEMENT_NAME);
-                String loopFormatted = getLoopPathD(half, side);
-                writer.writeAttribute(PATH_D_ATTRIBUTE, loopFormatted);
-                drawLoopEdgeInfo(writer, half, labelProvider.getEdgeInfos(graph, edge, side));
+                writer.writeAttribute(PATH_D_ATTRIBUTE, getLoopPathString(edge, side));
+                drawLoopEdgeInfo(writer, edge, side, labelProvider.getEdgeInfos(graph, edge, side));
             }
         }
         if (edge.getType().equals(BranchEdge.TWO_WT_EDGE)) {
-            draw2WtWinding(writer, half);
+            draw2WtWinding(writer, edge.getPoints(side));
         }
         writer.writeEndElement();
     }
 
-    private String getLoopPathD(List<Point> points, BranchEdge.Side side) {
-        double edgeStartAngle = getEdgeStartAngle(points);
+    private String getPolylinePointsString(BranchEdge edge, BranchEdge.Side side) {
+        return getPolylinePointsString(edge.getPoints(side));
+    }
+
+    private String getPolylinePointsString(ThreeWtEdge edge) {
+        return getPolylinePointsString(edge.getPoints());
+    }
+
+    private String getPolylinePointsString(List<Point> points) {
+        return points.stream()
+                .map(point -> getFormattedValue(point.getX()) + "," + getFormattedValue(point.getY()))
+                .collect(Collectors.joining(" "));
+    }
+
+    private String getLoopPathString(BranchEdge edge, BranchEdge.Side side) {
+        List<Point> points = edge.getPoints(side);
+        double edgeStartAngle = edge.getEdgeStartAngle(side);
         double controlsDist = 2 * (svgParameters.getEdgesForkLength() - svgParameters.getVoltageLevelCircleRadius());
         Point control1 = points.get(1).atDistance(controlsDist, edgeStartAngle);
 
@@ -410,17 +219,16 @@ public class SvgWriter {
     }
 
     private void drawThreeWtEdge(Graph graph, XMLStreamWriter writer, ThreeWtEdge edge) throws XMLStreamException {
+        if (!edge.isVisible()) {
+            return;
+        }
         writer.writeStartElement(GROUP_ELEMENT_NAME);
         writer.writeAttribute(ID_ATTRIBUTE, edge.getDiagramId());
         addStylesIfAny(writer, styleProvider.getEdgeStyleClasses(edge));
         insertName(writer, edge::getName);
         writer.writeEmptyElement(POLYLINE_ELEMENT_NAME);
-        List<Point> points = edge.getPoints();
-        String lineFormatted = points.stream()
-                .map(point -> getFormattedValue(point.getX()) + "," + getFormattedValue(point.getY()))
-                .collect(Collectors.joining(" "));
-        writer.writeAttribute(POINTS_ATTRIBUTE, lineFormatted);
-        drawEdgeInfo(writer, points, labelProvider.getEdgeInfos(graph, edge));
+        writer.writeAttribute(POINTS_ATTRIBUTE, getPolylinePointsString(edge));
+        drawThreeWtEdgeInfo(writer, edge, labelProvider.getEdgeInfos(graph, edge));
         writer.writeEndElement();
     }
 
@@ -468,12 +276,16 @@ public class SvgWriter {
         writer.writeAttribute(CIRCLE_RADIUS_ATTRIBUTE, getFormattedValue(svgParameters.getTransformerCircleRadius()));
     }
 
-    private void drawLoopEdgeInfo(XMLStreamWriter writer, List<Point> line, List<EdgeInfo> edgeInfos) throws XMLStreamException {
-        drawEdgeInfo(writer, edgeInfos, line.get(1), getEdgeStartAngle(line) - Math.PI / 2);
+    private void drawLoopEdgeInfo(XMLStreamWriter writer, BranchEdge edge, BranchEdge.Side side, List<EdgeInfo> edgeInfos) throws XMLStreamException {
+        drawEdgeInfo(writer, edgeInfos, edge.getPoints(side).get(1), edge.getEdgeStartAngle(side) - Math.PI / 2);
     }
 
-    private void drawEdgeInfo(XMLStreamWriter writer, List<Point> line, List<EdgeInfo> edgeInfos) throws XMLStreamException {
-        drawEdgeInfo(writer, edgeInfos, getArrowCenter(line), getEdgeEndYAxisAngle(line));
+    private void drawBranchEdgeInfo(XMLStreamWriter writer, BranchEdge edge, BranchEdge.Side side, List<EdgeInfo> edgeInfos) throws XMLStreamException {
+        drawEdgeInfo(writer, edgeInfos, getArrowCenter(edge.getPoints(side)), getEdgeEndYAxisAngle(edge.getPoints(side)));
+    }
+
+    private void drawThreeWtEdgeInfo(XMLStreamWriter writer, ThreeWtEdge edge, List<EdgeInfo> edgeInfos) throws XMLStreamException {
+        drawEdgeInfo(writer, edgeInfos, getArrowCenter(edge.getPoints()), getEdgeEndYAxisAngle(edge.getPoints()));
     }
 
     private void drawEdgeInfo(XMLStreamWriter writer, List<EdgeInfo> edgeInfos, Point infoCenter, double infoAngle) throws XMLStreamException {
@@ -531,12 +343,6 @@ public class SvgWriter {
         Point point1 = line.get(line.size() - 1);
         Point point0 = line.get(line.size() - 2);
         return Math.atan2(point1.getX() - point0.getX(), -(point1.getY() - point0.getY()));
-    }
-
-    private double getEdgeStartAngle(List<Point> line) {
-        Point point1 = line.get(1);
-        Point point0 = line.get(0);
-        return Math.atan2(point1.getY() - point0.getY(), point1.getX() - point0.getX());
     }
 
     private void draw2WtWinding(XMLStreamWriter writer, List<Point> half) throws XMLStreamException {
@@ -695,9 +501,9 @@ public class SvgWriter {
 
     private double getEdgeStartAngle(Edge edge, BranchEdge.Side side) {
         if (edge instanceof ThreeWtEdge) {
-            return getEdgeStartAngle(((ThreeWtEdge) edge).getPoints());
+            return ((ThreeWtEdge) edge).getEdgeStartAngle();
         } else if (edge instanceof BranchEdge) {
-            return getEdgeStartAngle(((BranchEdge) edge).getPoints(side));
+            return ((BranchEdge) edge).getEdgeStartAngle(side);
         }
         return 0;
     }
